requirements.txt
python-telegram-bot==20.2
aioredis==2.0.1


marshmallow==3.18.0
aiohttp-apispec==2.2.3